core Add C bindings and move WASM to it's separate folder as well
