+++ README.md
@@ -6,10 +6,16 @@ Reaching mastery takes decades, and only a few succeed. Why do people give up, a
 - qqself is also an app, an interactive journal that incorporates all the ideas from the book to mitigate the reasons why people give up and provides the motivation boost to persist in practical learning.
 
 ## core
-A shared Rust core handles journal entry parsing, encryption, all metric and identity calculations. All other services and apps heavily use it either directly or through WebAssembly.
+A shared Rust core handles journal entry parsing, encryption, all metric and identity calculations. All other services and apps heavily use it either directly or through WebAssembly or via C bindings.
+
+## core-bindings-c
+C bindings for the `core`. Used by the Swift `client-ios` and in the future by Kotlin `client-android`.
+
+## core-bindings-wasm
+WebAssembly bindings for the `core`. Used by the `client-web`.
 
 ## api-sync
-A Rust web service based on `actix-web` persists journal entries in AWS DynamoDB and allows multiple clients to synchronize. It operates with already encrypted payloads and cannot access user data.
+A Rust web service based on `actix-web` persists journal entries in `AWS DynamoDB` and allows multiple clients to synchronize. It operates with already encrypted payloads and cannot access user data.
 
 ## client-cli
 A Rust command-line client works with plain text journal files and allows you to have full control of the data. It also supports importing all data from `api-sync` to the plain text journal file and exporting the journal file back to the cloud. You are free to move from or to the synchronization server at any time.
