--- client-web/rsw.toml
@@ -1,8 +1,0 @@
-cli="yarn"
-
-[[crates]]
-name = "bridge"
-
-[crates.watch]
-# Encryption is CPU heavy. With dev profile it's very slow and unusable during development, switch to release
-profile = "release"
