+++ client-web/src/app/encryptionPool/pool.ts
@@ -1,4 +1,4 @@
-import { Keys } from "../../../bridge/pkg/qqself_client_web_bridge"
+import { Keys } from "../../../qqself_core"
 import { info } from "../../logger"
 import { isBrowser } from "../../utils"
 import { EncryptedEntry } from "../api"
