+++ client-web/src/app/encryptionPool/web-worker.ts
@@ -1,4 +1,4 @@
-import init, { initialize, Keys } from "../../../bridge/pkg/qqself_client_web_bridge"
+import init, { initialize, Keys } from "../../../qqself_core"
 import { InputType, OutputType, processMessage } from "./worker"
 
 let cachedKeys: Keys | null = null
