+++ client-web/src/app/init.ts
@@ -1,4 +1,4 @@
-import init, { initialize, Keys } from "../../bridge/pkg/qqself_client_web_bridge"
+import init, { initialize, Keys } from "../../qqself_core"
 import { info } from "../logger"
 import { isBrowser } from "../utils"
 import { getCredentials } from "./auth"
