+++ client-web/src/app/store.ts
@@ -1,4 +1,4 @@
-import { Keys, Views } from "../../bridge/pkg/qqself_client_web_bridge"
+import { Keys, Views } from "../../qqself_core"
 import { debug, info } from "../logger"
 import { APIProvider } from "./api"
 import * as Auth from "./auth"
