+++ client-web/src/ui/pages/login.ts
@@ -4,7 +4,7 @@ import "../controls/button"
 import { css, html, LitElement } from "lit"
 import { customElement, property, query, state } from "lit/decorators.js"
 
-import { Keys } from "../../../bridge/pkg/qqself_client_web_bridge"
+import { Keys } from "../../../qqself_core"
 import { Store } from "../../app/store"
 
 declare global {
