+++ client-web/vite.config.js
@@ -1,9 +1,7 @@
 /// <reference types="vitest" />
 import { defineConfig } from "vite"
-import ViteRsw from "vite-plugin-rsw"
 
 export default defineConfig({
-  plugins: [ViteRsw()],
   define: {
     "import.meta.vitest": "undefined",
   },
