+++ core-bindings-c/README.md
@@ -1,0 +1,3 @@
+# core-bindings-c
+
+Contains C bindings for the qqself core. We are using [uniffi](https://github.com/mozilla/uniffi-rs) to avoid writing bindings manually. Not everything is exposed from the core, but only the main features to support mobile client [client-ios](../client-ios).
