+++ core-bindings-c/build.rs
@@ -1,0 +1,3 @@
+fn main() {
+    uniffi::generate_scaffolding("./src/qqself.udl").unwrap();
+}
