+++ core-bindings-c/src/qqself.udl
@@ -1,0 +1,5 @@
+namespace qqself {
+  u32 add(u32 a, u32 b);
+  string hello();
+  string string_hash(string input);
+};
