+++ core-bindings-wasm/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "qqself-client-web-bridge"
+name = "qqself-core-bindings-wasm"
 version = "0.0.0"
 edition = "2021"
 build = "build.rs"
@@ -9,7 +9,7 @@ crate-type = ["cdylib", "rlib"]
 
 [dependencies]
 wasm-bindgen = "0.2.88"
-qqself-core = { path = "../../core", features = ["wasm"] }
+qqself-core = { path = "../core", features = ["wasm"] }
 console_error_panic_hook = "0.1.7"
 js-sys = { version = "0.3.65" }
