+++ core-bindings-wasm/README.md
@@ -1,0 +1,3 @@
+# core-bindings-wasm
+
+Contains WASM bindings for the qqself core. We are using [wasm-bindgen](https://github.com/rustwasm/wasm-bindgen) to generate bindings. Not everything is exposed from the core, but only the main features to support [client-web](../client-web/).
