+++ core-bindings-wasm/src/lib.rs
