+++ core-bindings-wasm/src/util.rs
